requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
telethon==1.36.0
google-generativeai==0.8.3
python-dotenv==1.0.1
cryptography==41.0.7
